client/requirements.txt
numpy==1.26.4
pillow==10.2.0
requests==2.32.3
pyaudio==0.2.14
opencv-python==4.9.0.80